backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pandas==2.2.1
numpy==1.25.2
python-jose==3.3.0
PyJWT==2.11.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
slowapi==0.1.8
networkx==3.2.1
scikit-learn==1.4.2
httpx==0.24.1
starlette==0.36.3
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
black==23.12.1
pylint==3.0.3
mypy==1.7.1
